profit calculator added for edit items page

+++ lib/datas/menu_data/write_data_menu.dart
@@ -108,7 +108,7 @@ class WriteData {
   }
 
 
-  Future<void> setExistingItemInMenu(String itemName, int newPriceMoney, int newPricePenny, List<String> indList, {int newProfit = 0}) async {
+  Future<void> setExistingItemInMenu(String itemName, int newPriceMoney, int newPricePenny, List<String> indList, {double newProfit = 0}) async {
     try {
       Map<String, dynamic>? rawData = await readData.getRawData();
